Pending test: Name hashing invalidation is insufficient because of subtyping
This test demonstrate that name hashing as implemented in sbt is
incorrect: When `AB2.scala` replaces `AB.scala` the hash of `x` does not
change, it only contains the syntactic signature of `B.x` which is
`this#A1`, the supertypes of the type of `B.x` are not recorded.
Therefore, `Test` is not recompiled and compilation succeeds when it
shouldn't.